+++ packages/in_app_purchase/in_app_purchase_android/android/src/main/java/io/flutter/plugins/inapppurchase/BillingClientFactoryImpl.java
@@ -6,7 +6,10 @@ package io.flutter.plugins.inapppurchase;
 
 import android.content.Context;
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import com.android.billingclient.api.BillingClient;
+import com.android.billingclient.api.UserChoiceBillingListener;
+import io.flutter.Log;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugins.inapppurchase.MethodCallHandlerImpl.BillingChoiceMode;
 
@@ -15,11 +18,34 @@ final class BillingClientFactoryImpl implements BillingClientFactory {
 
   @Override
   public BillingClient createBillingClient(
-      @NonNull Context context, @NonNull MethodChannel channel, int billingChoiceMode) {
+      @NonNull Context context,
+      @NonNull MethodChannel channel,
+      int billingChoiceMode,
+      @Nullable UserChoiceBillingListener userChoiceBillingListener) {
     BillingClient.Builder builder = BillingClient.newBuilder(context).enablePendingPurchases();
-    if (billingChoiceMode == BillingChoiceMode.ALTERNATIVE_BILLING_ONLY) {
-      // https://developer.android.com/google/play/billing/alternative/alternative-billing-without-user-choice-in-app
-      builder.enableAlternativeBillingOnly();
+    switch (billingChoiceMode) {
+      case BillingChoiceMode.ALTERNATIVE_BILLING_ONLY:
+        // https://developer.android.com/google/play/billing/alternative/alternative-billing-without-user-choice-in-app
+        builder.enableAlternativeBillingOnly();
+        break;
+      case BillingChoiceMode.USER_CHOICE_BILLING:
+        if (userChoiceBillingListener != null) {
+          // https://developer.android.com/google/play/billing/alternative/alternative-billing-with-user-choice-in-app
+          builder.enableUserChoiceBilling(userChoiceBillingListener);
+        } else {
+          Log.e(
+              "BillingClientFactoryImpl",
+              "userChoiceBillingListener null when USER_CHOICE_BILLING set. Defaulting to PLAY_BILLING_ONLY");
+        }
+        break;
+      case BillingChoiceMode.PLAY_BILLING_ONLY:
+        // Do nothing.
+        break;
+      default:
+        Log.e(
+            "BillingClientFactoryImpl",
+            "Unknown BillingChoiceMode " + billingChoiceMode + ", Defaulting to PLAY_BILLING_ONLY");
+        break;
     }
     return builder.setListener(new PluginPurchaseListener(channel)).build();
   }